Use uint8_t for pixel difference to account for representing a negative pixel difference using overflow

+++ pixel.h
@@ -12,7 +12,7 @@ typedef union {
 } pixel_t;
 
 typedef struct {
-	int16_t r, g, b, a;
+	uint8_t r, g, b, a;
 } pixel_difference_t;
 
 //
